Well-rounded Create Endpoint
The create endpoint will now fail under certain circumstances:

- Invalid employee input (fields or values)
- Employee input containing an emp_no

And upon finishing, it will return an object that represents what was added to the database with a code of 201, with proper error handling throughout if something went wrong.

This should amount to handling all the things I care about on the Create side.