Create Endpoint Test
This is the test that opened the giant can of worms. As it exists right now, it passes (which is good because the endpoint is functional) but it actually calls the endpoint and observes behavior. This means that it will only work on my local under very specific circumstances.

I want to fix this by mocking the behind-the-scenes stuff that is happening, but will need to do it later. As is, this commit was a two-day struggle just to figure out what was going on, so I'm closing my eyes and pretending its a success for the time being.