Work on the Update endpoint.
Along with other various items. Update endpoint no longer wants emp_no to work. Create endpoint is now actually returning the object it created instead of just mirroring the input with the new ID.

In addition, this update includes the application-dev.properties. . . except that it doesn't, given .gitignore. This change has no footprint in the repo beyond the gitignore, so I'm pointing it out here in a lot of words in case I need to find this commit later.