Update endpoint
This was tricky to come up with given the absence of ordering when it comes to getters on objects. I opted to construct the ordered arrayList with either null, or the eventual string that gets appended to the query.

This lets me iterate over the arraylist nicely, but is only enabled by appending the comma at the front of the stub being added to the query. Adding the comma at the front and not the end was the eureka in this situation, and it requires a boolean that gets toggled once any field has been added.