Error Handling
Using new Spring Annotations to hijack its native error handling to present cleaner and usable errors. This is somewhat incomplete, as I'm only overriding built-in methods that I've encountered thus far. I'm handling custom exceptions individually.

IntelliJ is telling me that the @AllArgsConstructor in ApiSubError is invalid (error shows up where it is used, in com.perficient.employees.error.ApiError. I don't think this is being called just yet, so it may actually be an error.

Overall this exercise was troublesome, but it's at about as stable a state as I can get it for now.