Env-specific properties files
Exposing passwords in .properties files is not the end of the world so long as SQL is locked down. 

I'm going to move toward including .properties files in the repo so hiding passwords doesn't result in hiding important configuration info. I suppose I'll find out pretty quick if this is in error.