DeployScript Argument Parsing
The idea here is that I'm going to use one deploy script for all deploys. By passing in arguments to specify: 

- environment (dev/prod/all)
- api (just employees for now)
- implementation (jdbc/jdbc_t/all)

Python's argparse has me spoiled, so there is some logic to obtain some semblance of reason if you're not familiar with the syntax.

Branching for checkpoint, and original functionality is commented away for reference.