Deploy Script Continued
Updated the Deploy script to handle file preparation. This is now done in the 'staging/' folder, rather than within the dev-env directory. 

Needed to update the vagrantfile and .gitignore to work with the new staging folder and remove references to dev-box.

I added 'clean' to the mvn package commands on account of JDBC_T failing inexplicably. Unsure if this is actually the fix, but it looks great for now.