Karate
New test IntelliJ project for Karate Integration tests. This will be run separately from the api projects, and will test the deployed endpoints.

As of right now, this test can be run, and can very likely be incorporated into the Jenkins pipeline. Checkpoint commit as I play around with some of the different features.