Cucumber HTML Reports Output
The goal here is to be able to run the Karate tests in such a way that outputs pretty HTML reports for browsing.

This took some doing, and ultimately I ran into a very bizarre problem with (seemingly) the location of the necessary properties and config files for this task. Fixed that, and it all came together. Google Drive for more details.