Deploy step
There's a lot going on with this commit, so much so that I expect many things to break. The plan from here is to attend to each of the breaking things one by one until it eventually gets the deploy onto prod (at which point I'm unconvinced it will work; prod is behind the times on configuration)

As I wrote this, I talked myself into making this a new branch. This way master will still complete its builds (unless the JAVA_HOME issue is still lingering). A branch makes most sense here.